수정 - refresh token이 계속 발급되는 문제 수정

+++ assignment-03-spring/src/main/java/com/kimilm/expert/security/jwt/JwtTokenUtils.java
@@ -16,7 +16,7 @@ public final class JwtTokenUtils {
     // JWT 토큰의 유효기간: 3일 (단위: seconds)
     private static final int JWT_TOKEN_VALID_SEC = 3 * DAY;
     // JWT 토큰의 유효기간: 3일 (단위: milliseconds)
-    private static final int JWT_TOKEN_VALID_MILLI_SEC = JWT_TOKEN_VALID_SEC * 1000;
+    private static final Long JWT_TOKEN_VALID_MILLI_SEC = JWT_TOKEN_VALID_SEC * 1000L;
 
     public static final String CLAIM_EXPIRED_DATE = "EXPIRED_DATE";
     public static final String CLAIM_USER_NAME = "USER_NAME";
